split webpack into dev/prod, split chunks
