+++ app-frontend/package.json
@@ -5,12 +5,10 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --mode development --open --hot",
-    "build": "webpack --mode production",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix --ext .js,.jsx ."
+    "start": "webpack-dev-server --config webpack.dev.js --open",
+    "build": "webpack --config webpack.prod.js"
   },
-  "author": "Kirill Stepanchuk <kirill.steapnchuk.96@mail.ru>",
+  "author": "Kirill Stepanchuk <kirill.stepanchuk.96@mail.ru>",
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.16.0",
@@ -19,15 +17,19 @@
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
     "babel-loader": "^8.2.3",
+    "clean-webpack-plugin": "^4.0.0",
+    "compression-webpack-plugin": "^9.0.1",
     "css-loader": "^6.5.1",
     "eslint": "^8.2.0",
     "eslint-plugin-react": "^7.26.1",
     "html-webpack-plugin": "^5.5.0",
     "redux-devtools-extension": "^2.13.9",
     "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.2.5",
     "webpack": "^5.62.2",
     "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.4.0"
+    "webpack-dev-server": "^4.4.0",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
     "@emotion/react": "^11.5.0",
@@ -47,6 +49,7 @@
     "redux": "^4.1.2",
     "redux-saga": "^1.1.3",
     "regenerator-runtime": "^0.13.9",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.3",
+    "webpack-bundle-analyzer": "^4.5.0"
   }
 }
