+++ app-frontend/src/store/reducers/worker.js
@@ -2,7 +2,6 @@ import initialState from '../initialState';
 import { LOAD_WORKER_DATA_SUCCESS, EDIT_WORKER_DATA } from '../constants';
 
 const worker = (state = initialState.worker, action) => {
-    console.log('state: ', state);
     switch (action.type) {
         case LOAD_WORKER_DATA_SUCCESS:
             return {
