+++ app-frontend/webpack.prod.js
@@ -1,0 +1,41 @@
+const path = require("path");
+const common = require("./webpack.common");
+const { merge } = require("webpack-merge");
+const TerserPlugin = require("terser-webpack-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+
+module.exports = merge(common, {
+    mode: "production",
+    output: {
+        filename: "[name].[hash:8].js",
+        path: path.resolve(__dirname, 'dist'),
+    },
+    optimization: {
+        splitChunks: {
+            chunks: 'all',
+            maxInitialRequests: Infinity,
+            minSize: 0,
+            cacheGroups: {
+                vendor: {
+                    test: /[\\/]node_modules[\\/]/,
+                    name: "vendor",
+                },
+            },
+        },
+        minimizer: [
+            new TerserPlugin()
+        ],
+    },
+    plugins: [
+        new HtmlWebpackPlugin({
+            template: "./src/index.html",
+            minify: {
+                removeAttributeQuotes: true,
+                collapseWhitespace: true,
+                removeComments: true,
+            }
+        }),
+        new BundleAnalyzerPlugin(),
+    ],
+});
